Use a single workflow per commit to upload from Buildomat to Trunk (#9249)
Iliana pointed out that the Trunk integration [fills the sidebar with
its
workflows](https://github.com/oxidecomputer/omicron/runs/52928547238),
hitting a limit in GitHub and thus not showing the Buildomat jobs. This
was because we were spawning an upload job for every Buildomat job.

This PR switches the upload job from running when a Buildomat job
finishes (with the `check_run` event) to when all Buildomat jobs finish
(with the `check_suite` event). Then whenever the job is started we
fetch all the check runs within the suite and call the script on all of
them.