Local storage 2/4: Add the new disk type
Add the ability to create LocalStorage type Disks.

From the API this means that the DiskCreate struct has been changed into
an enum, with variants for Crucible and LocalStorage. LocalStorage disks
will be attached and detached just like Crucible disks, but will
(currently) not support any of the other operations on a disk (snapshot,
import, etc).

Note that this is a breaking change to the Nexus external API - see
nexus/types/src/external_api/params.rs for more on this. Note too that
`content = "content"` was added to `InstanceDiskAttachment` because
otherwise the type wouldn't pass the openapi lint step.

A separate "local storage dataset allocation" row will be created during
instance allocation in the next PR, slicing a portion of the zpool's
local storage dataset. This will be delegated to the Propolis zone and
used as file backed block storage for an NVMe device. Using a zvol this
way was both measured to be the fastest in terms of 4k IOPs, and have
the least amount of standard deviation in those IOPs measurements,
according to fio runs.