initial Turin CPU platform (#9043)
this is so much simpler than codifying all the of the bits describing
all of the CPU surface area! what a breath of fresh air!

The feature selection here is the intersection of "PPR says it's there",
"useful for guests", "supported in byhve/propolis", and "doesn't seem
like we're painted into a corner if a future platform changes it." the
bits here are, also, a subset of what what I'd seen on a 9365 in a
Cosmo.

While byhve/Propolis would let guests turn on AutoIBRS, I haven't looked
at it in the context of guest OSes much at all (though they do _boot_
when told they're allowed to use AutoIBRS). UAI is in a similar boat but
I don't think anyone uses it. So both EFER features are hidden for the
time being.

Otherwise, as-is, I've booted Linux, Windows, OmniOS, and FreeBSD with
this profile and they seem fine. Linux for example omits mentioning
caches in `/proc/cpuinfo`, which make sense since I've avoided as much
cache topology information as I can here.. the reasoning for _that_ is
discussed more in RFD 314.