TQ: Implement network config replication
This code essentially re-implements the bootstore implementation of
early network config replication over sprockets channels.

The persistence of this state still lives in the bootstore, but will
eventually move to the `trust-quorum` crate once all customer systems
are running only trust quorum.

This code does not deal with how the switchover from bootstore to
to trust-quorum is made. That will come later with trust-quorum /
sled-agent integration.