[SCIM 4/4]: Draw the rest of the owl (#9276)
The majority of this PR fills out the CrdbScimProviderStore
implementation that the scim2-rs crate's Provider object will use to
implement SCIM where CRDB is the durable store for users and groups, and
adds a boat load of integration tests.

The notion of a silo user being "active" or not has also been added to
support users being deactivated by the SCIM client. Non-SCIM silo users
should not be affected.

---------

Co-authored-by: Mike Zeller <mike@oxidecomputer.com>