TQ: Support persisting state to ledger (#9310)
Builds on https://github.com/oxidecomputer/omicron/pull/9296

This commit persists state to a ledger, following the pattern used in
the bootstore. It's done this way because the `PersistentState` itself
is contained in the sans-io layer, but we must save it in the async task
layer. The sans-io layer shouldn't know how the state is persisted, just
that it is, and so we recreate the ledger for every time we write it.

A follow up will PR will deal with the early networking information
saved by the bootstore, and will be very similar.