TQ: Implement network config replication (#9317)
This builds on https://github.com/oxidecomputer/omicron/pull/9310

This code essentially re-implements the bootstore implementation of
early network config replication over sprockets channels.

The persistence code of this state still lives in the bootstore, but
will eventually move to the `trust-quorum` crate once all customer
systems are running only trust quorum.

This code does not deal with how the switchover from bootstore to to
trust-quorum is made. That will come later with trust-quorum /
sled-agent integration.