BlueprintBuilder: require caller to provide internal DNS subnet (#9250)
This is part of #9238: we'd like to pull `BlueprintResourceAllocator`'s
weird construction out of `BlueprintBuilder`, and this removes _one_ use
of it. When adding an internal DNS zone, the caller must now specify the
`DnsSubet` instead of `BlueprintBuilder` choosing internally.

To assist with this, adds
`BlueprintBuilder::available_internal_dns_subnets()`, which returns an
iterator of available subnets. I'm not sure this is the right interface;
will leave a comment inline with some other ideas.