Specify external DNS IPs in `PlanningInput` instead of deriving them from the parent blueprint (#9291)
This is built on top of #9250, and is related to both #9238 and #8949:

* Trims down some of the `BlueprintResourceAllocator` internals
* Adds an explicit `ExternalIpPolicy` type; today it holds a couple of
IP pools and a set of external DNS IPs, but it gives us a cleaner place
to integrate all the IP pool work that @bnaecker is doing with
Reconfigurator

Since we don't actually have external DNS IPs in a policy anywhere, this
adds `DataStore::external_dns_external_ips_specified_by_rack_setup()`
(which is implemented as "load the current target blueprint and scan it
for external DNS zones"; i.e., exactly what the builder was doing
before). We should eventually delete this method once it's possible for
an operator to reconfigure the external DNS IPs, but this lets the
planning input build up the policy as though we already had this
information available, more or less. Other than that new method,
_almost_ all the changes in this PR are to tests; the non-test changes
are pretty minimal.