Update `gateway-messages` and bundled `faux-mgs`
In particular, this picks up
https://github.com/oxidecomputer/management-gateway-service/pull/451 and
https://github.com/oxidecomputer/management-gateway-service/pull/452,
which we'd like for debugging hung sled issues.

To make this easier to backport to the R17 release branch, it makes
minimal changes: requesting these new types of component details from
MGS proper will return a 400 (albeit one that probably contains all the
details we actually want in the error message!), because returning a 200
would require revving the MGS OpenAPI. The control plane has no logic to
collect this information yet, so I think this is fine - it'll be
available to support via `faux-mgs` as we'd like, and I'll file an issue
for adding proper MGS support at our convenience.