Local storage [1/4]: add local_storage dataset (#9264)
To implement RFD 590, Propolis will use a delegated ZVOL as a file
backed NVMe device. This PR creates a new "local_storage" dataset that
can be used for this purpose on every zpool in the rack. This is created
under the `crypt` parent so that the delegated storage is encrypted at
rest.

This is done with the planner instead of adding to U2_EXPECTED_DATASETS,
and this required changing a bunch of tests that were assuming a
specific number of datasets per pool, along with a whole whack of tests
that use expectorate.