nexus: make oximeter database pool size configurable.
We use qorb as the connection pooler for ClickHouse within oximeter, and
inherit the default pool max_slots of 16. We may be saturating that pool size
for some OxQL workloads, such as the proposed otel receiver, which runs many
tiny queries in parallel against ClickHouse. This patch makes the pool size
configurable, so that we can adjust it if necessary.