[nexus] fault management situation reports (#9320)
[RFD 603] proposes the _fault management situation report_, or _sitrep_,
as the central data structure for the control plane's fault management
subsystem. The design, which is discussed in much greater detail in that
RFD, draws a lot of inspiration from the blueprint data structure in the
Reconfigurator. Sitreps are generated by the planning phase in a
plan-execute pattern. At any time, a single sitrep is considered
_current_. Updating the control plane's understanding of the state of
the system based on new inputs is done by a new planning step based on
the current sitrep along with other inputs, and produces a new sitrep
with the current sitrep as its _parent_. A sitrep may then be added to
the version history of current sitreps if (and only if) its parent
sitrep is still the current sitrep (i.e. the highest version number
currently stored in the sitrep history). This ensures that there is a
single sequentially consistent history of sitreps. Sitreps generated
based on outdated inputs --- due to multiple Nexuses generating them
concurrently, or a Nexus operating on state that is no longer up to date
---may not be made current, and are discarded.

This branch adds the foundation of the sitrep subsystem. In particular,
it includes the following:

- Database schemas for the `fm_sitrep` table, which stores metadata for
sitreps, and the `fm_sitrep_history` table, which stores the version
history
- Models and `nexus_types` types for the same
- Database queries for reading the current sitrep version, reading a
sitrep by its ID, and for inserting sitreps, including the "compare and
swap" CTE that ensures new versions may only be inserted if they descend
directly from the current sitrep
- A `fm_sitrep_loader` task that loads the latest sitrep version and
publishes it over a `tokio::sync::watch` channel (which is not presently
consumed by other code)
- OMDB commands for looking at sitreps

Right now, a sitrep only contains its top-level metadata. Other tables
for storing parts of the sitrep, such as
[cases](https://rfd.shared.oxide.computer/rfd/0603#_cases) and records
for [updating
Problems](https://rfd.shared.oxide.computer/rfd/0603#_problems), will be
added later as more of the control plane fault management subsystem is
implemented. Currently, no sitreps are ever created outside of tests, so
this code won't really _do_ anything yet. But, it's an important
foundation for the ret of the FM work, so I wanted to get it up for
review as soon as possible.

[RFD 603]: https://rfd.shared.oxide.computer/rfd/0603