Add `Blippy::check_against_planning_input()` (#9356)
This PR adds a new method to (optionally) check a blueprint against a
given planning input. The only check included in this PR is a
line-for-line port of the builder's
`ensure_input_networking_records_appear_in_parent_blueprint()` function
(which I'm removing from the builder in #9347). I updated most of the
`Blippy` call sites to use this method; I skipped the one in an RSS test
because it has no planning input available - RSS just generates a raw
blueprint to hand off to Nexus.