oximeter: simplify histogram response. (#9366)
Simplify oximeter histogram responses to show the estimates for each
quantile, rather than details of the P² algorithm.

Part of #9345.

Notes:
* If we're only showing the quantile estimates and not other details of
the algorithm, we could also simplify our clickhouse queries for
histograms. We can add this here, or get to it separately later on.
* This is technically a breaking change, in the event that customers are
using quantile fields in histograms—although I'm inclined to doubt they
are, because I don't think we document how to use those fields. Not sure
if this is a problem, but wanted to note it.