[nexus] garbage collect orphaned FM sitreps (#9335)
When a Nexus attempts to commit a new fault management situation report
to the sitrep history but fails to do so because another sitrep with the
same parent has already been inserted, that sitrep is said to be
_orphaned_. Records pertaining to it are left behind in the database,
but it will not be accessed by the rest of the system. Thus, we must
occasionally garbage-collect such sitreps. This branch adds a background
task for doing so.

Depends on #9320

Co-authored-by: Sean Klein <sean@oxide.computer>