Move `ClickhouseClusterConfig` construction from `BlueprintBuilder` to planner (#9370)
This is staged on top of #9347 and is a part of #9238. In particular, by
moving this from the builder to the planner, we can remove the builder's
need for an inventory collection entirely.

There should be no functional changes here; it's just moving code around
(with some reformatting to break it up a bit) and then removing the
`collection` argument from all the callers of
`BlueprintBuilder::new_based_on()`.