test-utils: surface dpd errors. (#9379)
At the moment, DendriteInstance::start in omicron-test-utils doesn't
catch errors in dpd. We spawn a child process, wait up to the timeout
interval to discover the dpd port from its logs, and fail if we can't
discover the port. But if dpd exits quickly, we still wait the timeout
interval and swallow any errors. This patch captures the stderr of the
dpd child process, checks it for an early exit, and prints the details
if dpd crashes.

This was bugging me when I was having trouble getting simulated omicron
to start due to my PATH pointing to the wrong dpd executable. h/t
@david-crespo.