`BlueprintBuilder`: don't hold on to `PlanningInput` (#9374)
This builds on top of #9370 and is a big piece of #9238: as of this PR,
`BlueprintBuilder` no longer holds a `PlanningInput`, which should
reduce confusion on where a given bit of logic should live. Anything
that needs the planning input to be decided now _must_ be implemented in
the planner.

`BlueprintBuilder::new_based_on(..)` still takes a `PlanningInput`
argument, which I don't love but need to do some more work to figure out
how (if?) to address. It uses it for a few things; I think I'd group
them as:

1. Reassembling information about sleds that we knew in previous
iterations but that aren't stored in the blueprint. (Specifically: the
`BaseboardId` and subnet of each active sleds. This is the one that
seems trickiest to tackle.)
2. Adding new, empty `SledEditor`s for commissioned sleds that are
present in `PlanningInput` but not the parent blueprint. (I think I'd
make the case that `new_based_on()` shouldn't do this at all, and the
planner should do it?)
3. Making a copy of a few fields that are present in the blueprint, but
that are updated each time the planner runs based on potentially-newer
values in `PlanningInput`. (Specifically: the cockroachdb fingerprint
and the internal/external DNS generations. I think it would be fine for
`BlueprintBuilder` to only copy the parent blueprint, and provide
setters that the planner could use to update these to newer values.)

There's one hopefully-small-and-fine logical change in this PR that I'll
point out below with some extra details.