sled_upsert(): also insert a corresponding `hw_baseboard_id` row
The `sled` table has always stored a `part_number`, `serial_number`
pair. For inventory and reconfigurator, that pair uniquely identifies a
sled. Those subsystems use the `hw_baseboard_id` table to normalize that
pair to a single ID, but we haven't done the work to go back and
normalize `sled` (a nontrivial chunk of work!). This PR does a half-step
in that direction: when inserting a sled, we can at least guarantee that
a matching `hw_baseboard_id` row _exists_, even if `sled` itself doesn't
reference it.