Better management of disconnecting peers
Rather than send a message when a peer disconnects and then handle that
message asynchronously, we now wait for the task itself to exit and then
return a `DisconnectedPeer` from `ConnectionManager::step`.

We only return the `PeerId` for the `DisconnectedPeer` if it is
still the existing `established connection` for the given peer and
it hasn't been replaced by a newer connection. This prevents calling
`Node::on_disconnect` for the stale connection when it might have
already received an `on_connect` call for the new connection.