Swap how to refer to the new instance ensure body
Instead of using a v7 module to refer to the new instance ensure body,
use a v6 module to refer to the old one - eventually we'll want to
deprecate and remove the old instance ensure body type, and it's easier
to refer to the old type in the few places that it's required to.

This commit also removes `multicast_cross_version_works` because I don't
think it's necessary now that we're sending (and routing based on!) the
api version header.