Simplify the versioning of old sled-agent types
This adds only a v8 into the public sled-agent-types crate with the
previous version. It adds infallible conversions from older API types to
and from v8, rather than changing those conversions to be directly to
the latest type. Also call the "next" version of the actual API handler
too, rather than jumping to the latest.