Make `UnstableReconfiguratorState::target_blueprint` non-optional
`UnstableReconfiguratorState`s constructed from real systems always
supplied a target blueprint, because real systems are required to always
have one. We'd kept this as optional to allow for `reconfigurator-cli`'s
simulated system to not have a blueprint at all, but I'd like to make
other changes related to the `PlanningInput` that trip over this
difference.

This PR addresses the simulator side by having it claim to have a
blueprint with the nil ID and containing nothing any time it's asked for
the current list of blueprints or current target blueprint and one
hasn't otherwise been set (i.e., by loading an example system or an
existing serialized state).