Add dual-stack IP config for internal NICs (#9389)
- Add dual-stack VPC private address configuration type and include it
in the shared NetworkInterface type.
- Add database model support for reading / writing the dual-stack NIC
type to the database, handling serialization into optional fields.
- Update all the callsites to handle the new dual-stack-aware type.
- Add a bunch of conversions for the APIs which rely on that new type,
of which there are many. This also adds the conversions and older types
into the `sled-agent-types` crate, so they can be used in a few places
that don't directly depend on the `sled-agent-api` crate itself, notably
reconfigurator and `nexus-inventory`.
- Update the sled-agent reconciler to deserialize previous versions of
its sled-configuration ledgers, convert them, and write them out again
as the new versions.
- Updates the OPTE `PortManager` type with the new dual-stack support.
This is only for private IP addresses, though. We still need some work
to support OPTE ports with dual stack external addresses. This is about
half of #9247, the VPC-private part.
- Closes #9246