`BlueprintSledConfig`: add sled subnet (#9416)
Sled subnets are chosen when the sled is added to the cluster (either by
RSS or by the "add sled" process), and do not change for the lifetime of
the sled's membership. This PR adds that subnet to
`BlueprintSledConfig`, removing one of the remaining bits where
`BlueprintBuilder::new_based_on()` needs information other than the
parent blueprint.