[ls-apis] check that only omicron-sled-agent depends on nexus-lockstep-client (#9413)
Sled Agent depends on nexus-lockstep-client via RSS, and we now ensure
that nothing else that's part of one of the deployment units depends on
it. (We don't treat omdb as part of a deployment unit because it doesn't
participate in upgrade in the typical sense.)

Also add an expectorate test for the output of `ls-apis check`, similar
to the existing one for `ls-apis apis`.

Closes #9057.