Temporarily check for expected hubris ordering
Omcrion #9437 fixed a bug that could cause an incorrect image
to be selected for update. Due to some luck with ordering, this
bug should not affect gimlet sleds so long as we have gimlet
images show up before cosmo images. Add some checks at build
time to make sure this is the case until we have upgraded pass
this bug.