Include /var/adm/messages in archived zone logs
System messages on illumos are logged to `/var/adm/messages`. In the
event of a hardware or kernel issue, an archive of these logs can be
for understanding the nature of the failure.

Currently we archive all Oxide service logs from the global zone and any
zones we create, but not `/var/adm/messages`. Update the archiver to
include the latter as well.